airflow/requirements.txt
pyspark==3.5.2
scikit-learn==1.5.2
mlflow==2.16.2
pandas==2.2.3
apache-airflow-providers-mongo
seaborn
effector @ git+https://github.com/givasile/effector.git@dfdf4d4b59819219fbee2abe9ac3b2b7ac91f1ba
PyGithub
pymongo
gitpython
nbformat

